TMDM-11781 Refactor the code, and run model/view validation within a job.
