+++ main/plugins/org.talend.mdm.repository/src/main/java/org/talend/mdm/repository/ui/wizards/imports/handlers/DataModelImportHandler.java
@@ -12,19 +12,12 @@
 // ============================================================================
 package org.talend.mdm.repository.ui.wizards.imports.handlers;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.swt.widgets.Display;
 import org.talend.commons.exception.PersistenceException;
-import org.talend.core.GlobalServiceRegister;
 import org.talend.core.model.properties.Item;
 import org.talend.core.model.repository.IRepositoryViewObject;
-import org.talend.mdm.repository.core.service.IModelValidationService;
+import org.talend.mdm.repository.model.mdmproperties.WSDataModelItem;
 import org.talend.mdm.repository.ui.wizards.imports.OperatorUpdatorProvider;
-import org.talend.mdm.repository.utils.RepositoryResourceUtil;
-import org.talend.repository.items.importexport.handlers.imports.IImportResourcesHandler;
 import org.talend.repository.items.importexport.handlers.model.ImportItem;
 import org.talend.repository.items.importexport.manager.ResourcesManager;
 
@@ -34,7 +27,7 @@ import com.amalto.workbench.exadapter.ExAdapterManager;
  * created by HHB on 2015-5-4 Detailled comment
  *
  */
-public class DataModelImportHandler extends CommonMdmImportHandler implements IImportResourcesHandler {
+public class DataModelImportHandler extends CommonMdmImportHandler {
 
     private IDataModelImportHandlerExAdapter exAdapter;
 
@@ -59,43 +52,8 @@ public class DataModelImportHandler extends CommonMdmImportHandler implements II
     }
 
     @Override
-    public void postImport(IProgressMonitor monitor, ResourcesManager resManager, ImportItem[] importedItemRecords) {
-        validateModelAsyc(importedItemRecords);
-    }
-
-    @SuppressWarnings("unused")
-    private void validateModelAsyc(ImportItem[] importedItemRecords) {
-        if (GlobalServiceRegister.getDefault().isServiceRegistered(IModelValidationService.class)) {
-            List<IRepositoryViewObject> viewObjs = new ArrayList<IRepositoryViewObject>();
-            for (ImportItem item : importedItemRecords) {
-                IRepositoryViewObject viewObj = RepositoryResourceUtil.findViewObjectById(item.getProperty().getId());
-                viewObjs.add(viewObj);
-            }
-
-            Display.getDefault().asyncExec(new Runnable() {
-
-                @Override
-                public void run() {
-                    IModelValidationService service = (IModelValidationService) GlobalServiceRegister.getDefault()
-                            .getService(IModelValidationService.class);
-                    service.validate(viewObjs, IModelValidationService.VALIDATE_IMMEDIATE, true);
-                }
-            });
-        }
-    }
-
-    /////////////////////////////////
-    @Override
-    public void prePopulate(IProgressMonitor monitor, ResourcesManager resManager) {
-    }
-
-    @Override
-    public void postPopulate(IProgressMonitor monitor, ResourcesManager resManager, ImportItem[] populatedItemRecords) {
-    }
-
-    @Override
-    public void preImport(IProgressMonitor monitor, ResourcesManager resManager, ImportItem[] checkedItemRecords,
-            ImportItem[] allImportItemRecords) {
+    public boolean todoValidate(ImportItem item) {
+        return item.getItem() instanceof WSDataModelItem;
     }
 
 }
