+++ main/plugins/org.talend.mdm.repository/src/main/java/org/talend/mdm/repository/ui/wizards/imports/handlers/ViewProcessImportHandler.java
@@ -12,32 +12,21 @@
 // ============================================================================
 package org.talend.mdm.repository.ui.wizards.imports.handlers;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.eclipse.core.runtime.IPath;
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.swt.widgets.Display;
-import org.talend.core.GlobalServiceRegister;
 import org.talend.core.model.properties.Item;
-import org.talend.core.model.repository.IRepositoryViewObject;
 import org.talend.mdm.repository.core.impl.transformerV2.ITransformerV2NodeConsDef;
 import org.talend.mdm.repository.core.impl.view.IViewNodeConstDef;
-import org.talend.mdm.repository.core.service.IModelValidationService;
 import org.talend.mdm.repository.model.mdmproperties.WSTransformerV2Item;
 import org.talend.mdm.repository.model.mdmproperties.WSViewItem;
 import org.talend.mdm.repository.ui.wizards.imports.OperatorUpdatorProvider;
-import org.talend.mdm.repository.utils.RepositoryResourceUtil;
 import org.talend.mdm.repository.utils.RepositoryTransformUtil;
-import org.talend.repository.items.importexport.handlers.imports.IImportResourcesHandler;
 import org.talend.repository.items.importexport.handlers.model.ImportItem;
-import org.talend.repository.items.importexport.manager.ResourcesManager;
 
 /**
  * created by HHB on 2015年5月4日 Detailled comment
  *
  */
-public class ViewProcessImportHandler extends CommonMdmImportHandler implements IImportResourcesHandler {
+public class ViewProcessImportHandler extends CommonMdmImportHandler {
 
     @Override
     protected void beforeCreatingItem(ImportItem selectedImportItem) {
@@ -99,44 +88,8 @@ public class ViewProcessImportHandler extends CommonMdmImportHandler implements 
     }
 
     @Override
-    public void postImport(IProgressMonitor monitor, ResourcesManager resManager, ImportItem[] importedItemRecords) {
-        if(importedItemRecords.length > 0 && importedItemRecords[0].getItem() instanceof WSViewItem) {
-            validateModelAsyc(importedItemRecords);
-        }
+    public boolean todoValidate(ImportItem item) {
+        return item.getItem() instanceof WSViewItem;
     }
 
-    @SuppressWarnings("unused")
-    private void validateModelAsyc(ImportItem[] importedItemRecords) {
-        if (GlobalServiceRegister.getDefault().isServiceRegistered(IModelValidationService.class)) {
-            List<IRepositoryViewObject> viewObjs = new ArrayList<IRepositoryViewObject>();
-            for (ImportItem item : importedItemRecords) {
-                IRepositoryViewObject viewObj = RepositoryResourceUtil.findViewObjectById(item.getProperty().getId());
-                viewObjs.add(viewObj);
-            }
-
-            Display.getDefault().asyncExec(new Runnable() {
-
-                @Override
-                public void run() {
-                    IModelValidationService service = (IModelValidationService) GlobalServiceRegister.getDefault()
-                            .getService(IModelValidationService.class);
-                    service.validate(viewObjs, IModelValidationService.VALIDATE_IMMEDIATE, true);
-                }
-            });
-        }
-    }
-
-    /////////////////////////
-    @Override
-    public void prePopulate(IProgressMonitor monitor, ResourcesManager resManager) {
-    }
-
-    @Override
-    public void postPopulate(IProgressMonitor monitor, ResourcesManager resManager, ImportItem[] populatedItemRecords) {
-    }
-
-    @Override
-    public void preImport(IProgressMonitor monitor, ResourcesManager resManager, ImportItem[] checkedItemRecords,
-            ImportItem[] allImportItemRecords) {
-    }
 }
